Sphinx: custom HTML generation example.
This is the beginning of an experiment on how to tweak and modify
the HTML Translator to generate the desired output.

So far, progress is good. However, it now needs a specialist node
to be registered, that can be used to group or otherwise delineate
where each group of related functions begin and end.

Hooking into `depart_desc_signature_line` only grabs the signature
for a single `{cpp:function}` definition, but the overall design
looks sound at this stage.